refactor: transferred code from pages to component
